pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-patterns"
version = "0.1.0"
description = "Reusable building blocks for production LangGraph agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Subramonian Inian" }]
dependencies = [
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5.0"]

[tool.setuptools.packages.find]
include = ["patterns*", "shared*"]

[tool.ruff]
line-length = 100
target-version = "py311"
